Allowed actor meshes to be shared with pointer system

+++ src/libvolume/engine/realm.h
@@ -18,7 +18,7 @@ namespace LibVolume
 		class Realm
 		{
 			public:
-				double time;
+				long long time;
 
 				Camera camera;
